Update our bootstrapping processing.
I found the existing tool nuvo/docker-cqlsh to be lacking some key feautres and/or having some subpar DX. For example:
- Using the "show version" command to check server status led to hard-to-understand error messages
- There was no way to provide a username/password to the cqlsh command
- Lacking best practices in the entrypoint.sh script leading to some strange cases where the connection command could fail but the script continues onward to attempt to execute CQL expressions

The nuvo/docker-cqlsh code base was rather small. So I rebuilt it in `cassandra/docker/cql_runner`. Some key changes here include:
1. Base Docker layer is the `cassandra` image (with Python 3) versus the python2-alpine image.
2. Require versions to be specified and allow username/password values to be piped through.
3. Smarter entrypoint script that fails properly.

I also added a README file to describe steps to bootstrap Cassandra locally. We will be adding more scripts in the future but that should be easy now that we have these initial steps figured out. The salient change here is that I ended up splitting the scripts into two folders, the bootstrap_scripts and scripts folder. This is to make the distinction that some things need to run as the default cassandra user, and once that initial bootstrap has occurred all other scripts can be run as your superuser role.