Set up proper emualtor and really simplify tools.
Instead of having all these shell scripts and steps to boot up Cassandra, I wanted to make it incredibly easy (and idiot proof) to boot the database up. So I wrapped up most of the steps inside a multi-stage docker build file, and built out the beginnings of an emulator with a docker compose file (compose.yaml).

The Cassandra database can now be spun up locally with a single command: `make em`. This comes from the new `Makefile` which runs the appropriate docker compose command.