Remove cassandra from the services folder.
It's not really a service in the sense that we'll be running it ourselves. We run it locally while developing/testing, but in dev and production I suspect we'll use a managed service.

The main contents of our local Cassandra service is in the emulator folder now. This is where I would expect local emulated services to live.