Fix services race condition during spin up.
Some services wait for Cassandra to be healthy before trying to connect. But they will attempt to connect a split second after the healthcheck passes, and even though the healthcheck confirms the keyspace was created the Cassandra client is returning errors of the keyspace missing.

Adding some retry mechanisms to the connection function in the db package helps prevent container failure.