Huge refactor of Kafka package.
Refactored the package to aid in writing the first test. Heavy use of mocks and dependency injection.

This change also moves the concept of Kafka down a level. The package is no longer a "kafka package", but a "message queue" package. And the implementation of the various interfaces could be Kafka related, or for the benefit of tests, a mock struct.